* Added a variation of finding sub array with given sum - handles negative numbers - with constant space usage. But seems algorithm is wrong.

+++ src/arrays/misc/sub_array_with_given_sum_v1.rs
@@ -8,7 +8,7 @@ pub fn find_from_positive_numbers(arr: &[i32], target: i32) -> (i32, i32) {
     for index in 1..arr.len() {
         sum += arr[index];
 
-        while (sum > target) && (start_index < index as i32) {
+        while (sum > target) && ((start_index as usize) < index) {
             sum -= arr[start_index as usize];
             start_index += 1;
         }
